wip: move route source to normalization
