+++ src/routing/normalization/resolve-route-options.ts
